+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Added
+- Warning to `utils.uvcalibrate` when data and cal x_orientation are mismatched.
 - Better re-initialization of UVParameters on UVBase objects when using pickle. Improves compatibility of UVBase objects with MPI.
 - Added option to apply a Van Vleck correction to mwa_corr_fits files.